ager-auth-app/requirements.txt
Django==4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
gunicorn==20.1.0